Move ActivityReports::AntenneMatches files together
