--- app/jobs/activity_reports/antenne_matches/generate_job.rb
@@ -1,8 +1,0 @@
-class ActivityReports::AntenneMatches::GenerateJob < ApplicationJob
-  queue_as :low_priority
-
-  def perform(antenne_id)
-    antenne = Antenne.find(antenne_id)
-    ActivityReports::Generate::AntenneMatches.new(antenne).call
-  end
-end
