--- spec/jobs/activity_reports/antenne_matches/generate_job_spec.rb
@@ -1,11 +1,0 @@
-require 'rails_helper'
-RSpec.describe ActivityReports::AntenneMatches::GenerateJob do
-  describe 'enqueue a job with an antenne' do
-    let(:antenne) { create(:antenne) }
-
-    it do
-      described_class.perform_later(antenne.id)
-      assert_enqueued_with(job: described_class, args: [antenne.id])
-    end
-  end
-end
