+++ spec/services/activity_reports/antenne_matches_spec.rb
@@ -1,5 +1,14 @@
 require 'rails_helper'
-describe ActivityReports::Generate::AntenneMatches do
+describe ActivityReports::AntenneMatches do
+  describe ActivityReports::AntenneMatches::Generate do
+    let(:antenne) { create(:antenne) }
+
+    it do
+      described_class.perform_later(antenne)
+      assert_enqueued_with(job: described_class, args: [antenne])
+    end
+  end
+
   describe 'generate_files' do
     let(:antenne) { create :antenne }
     let!(:expert) { create :expert_with_users, antenne: antenne }
